+++ lib/api-helpers.js
@@ -1,4 +1,6 @@
 
+
+// function for stripping down user object. 
 export const extractUser = req => {
   if (!req.user) return null;
   // take only needed user fields to avoid sensitive ones (such as password)
@@ -10,4 +12,6 @@ export const extractUser = req => {
   };
 }
 
+
+// Fetch wrapper for SWR/useSWR hook.
 export const fetcher = (...args) => fetch(...args).then(res => res.json());
