+++ lib/dateHelpers.js
