+++ lib/dbConnect.js
@@ -2,7 +2,7 @@ import mongoose from 'mongoose';
 
 const connection = {}; /* for caching db connection*/
 
-async function dbConnect(logComment) {
+export async function dbConnect(logComment) {
   // Debugging thing.
   if (logComment) {
     console.log(`dbConnect: ${logComment}`);
@@ -24,4 +24,3 @@ async function dbConnect(logComment) {
   connection.isConnected = db.connections[0].readyState;
 }
 
-export default dbConnect;
