+++ lib/passport-config.js
@@ -34,14 +34,18 @@ passport.use(new LocalStrategy(
   { passReqToCallback: true, usernameField: "email" },
   // callback function:
   async (req, email, password, done) => {
+
     // RegExp for insensitive case, and exact matching
     const regex = new RegExp(`^${email}$`, "i");
 
+    // Find user by email, and make sure to get the password
+    // PW doesn't get sent normally, on purpose. Set in our model
     const user = await User.findOne({ email: { $regex: regex } }).select("+password");
 
     // Verify user exists and entered password matches hashed password on file
     if (user && (await compare(password, user.password))) {
-
+      // TODO I think that serialize user gets called immediately after this, and strips all but the id.
+      // If that's incorrect, we gotta get that password off of here before we send that out.
       return done(null, user);
     } else {
       return done(null, false);
